refactor: attestation support
