+++ app/__tests__/screens/Developer.test.tsx
@@ -3,7 +3,7 @@ import { render } from '@testing-library/react-native'
 import React from 'react'
 
 import Developer from '../../src/screens/Developer'
-import { AttestationProvider } from '../../src/services/attestation'
+import { AttestationProvider } from '../../src/hooks/useAttestation'
 import { initialState, reducer } from '../../src/store'
 
 const mockNavigation = jest.fn()
