+++ app/src/constants.ts
@@ -2,3 +2,10 @@ export const autoDisableRemoteLoggingIntervalInMinutes = 60
 export const surveyMonkeyUrl = 'https://www.surveymonkey.com/r/7BMHJL8'
 export const surveyMonkeyExitUrl = 'https://www.surveymonkey.com/survey-thanks'
 export const hitSlop = { top: 44, bottom: 44, left: 44, right: 44 }
+export const attestationCredDefIds = [
+  'NXp6XcGeCR2MviWuY51Dva:3:CL:33557:bcwallet',
+  'NXp6XcGeCR2MviWuY51Dva:3:CL:33557:bcwallet_dev',
+  'RycQpZ9b4NaXuT5ZGjXkUE:3:CL:120:bcwallet',
+  'RycQpZ9b4NaXuT5ZGjXkUE:3:CL:120:bcwallet_test',
+  'XqaRXJt4sXE6TRpfGpVbGw:3:CL:655:bcwallet',
+]
