+++ app/src/index.ts
@@ -19,6 +19,7 @@ import HomeFooterView from './components/HomeFooterView'
 import HomeHeaderView from './components/HomeHeaderView'
 import { setup, activate, deactivate, status } from './helpers/PushNotificationsHelper'
 import { useNotifications } from './hooks/notifications'
+import { useAttestation } from './hooks/useAttestation'
 import en from './localization/en'
 import fr from './localization/fr'
 import ptBr from './localization/pt-br'
@@ -28,7 +29,6 @@ import PersonCredential from './screens/PersonCredential'
 import Preface from './screens/Preface'
 import Splash from './screens/Splash'
 import Terms from './screens/Terms'
-import { useAttestation } from './services/attestation'
 import { BCDispatchAction } from './store'
 import { defaultTheme as theme } from './theme'
